tests/e2e/pyproject.toml
[tool.poetry]
name = "schoolsyst-webapp-e2e-tests"
version = "0.1.0"
description = "E2E Tests for app.schoolsyst.com, made with Helium."
authors = ["ewen-lbh <ewen.lebihan7@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
helium = "^3.0.2"
python-dotenv = "^0.12.0"
behave = "^1.2.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
